ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.200
pillow==10.1.0
numpy==1.24.3
python-multipart==0.0.6
pydantic==2.5.0
